Sync up with the latest cvs (0.65b2)
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@431 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Weapon.cpp
@@ -173,7 +173,7 @@ void CWeapon::Update()
 			if(weaponDef->stockpile)
 				reloadStatus=gs->frameNum+60;
 			else
-				reloadStatus=gs->frameNum+reloadTime/owner->reloadSpeed;
+				reloadStatus=gs->frameNum+(int)(reloadTime/owner->reloadSpeed);
 
 			salvoLeft=salvoSize;
 			nextSalvo=gs->frameNum;
